+++ jaorm-test/src/test/java/io/github/ulisse1996/jaorm/integration/test/query/CityDAO.java
@@ -1,8 +1,15 @@
 package io.github.ulisse1996.jaorm.integration.test.query;
 
-import io.github.ulisse1996.jaorm.integration.test.entity.City;
 import io.github.ulisse1996.jaorm.BaseDao;
 import io.github.ulisse1996.jaorm.annotation.Dao;
+import io.github.ulisse1996.jaorm.annotation.Query;
+import io.github.ulisse1996.jaorm.integration.test.entity.City;
+
+import java.util.stream.Stream;
 
 @Dao
-public interface CityDAO extends BaseDao<City> {}
+public interface CityDAO extends BaseDao<City> {
+
+    @Query(sql = "SELECT * FROM CITY WHERE CITY_ID > ?")
+    Stream<City> getCityHigherThanId(Integer id);
+}