+++ jaorm-processor/src/test/resources/relationship/EnumConverter.java
@@ -1,0 +1,18 @@
+package io.test;
+
+import io.github.ulisse1996.jaorm.entity.converter.ValueConverter;
+
+public class EnumConverter implements ValueConverter<String, CustomEnum> {
+
+    public static final EnumConverter INSTANCE = new EnumConverter();
+
+    @Override
+    public CustomEnum fromSql(String val) {
+        return CustomEnum.valueOf(val);
+    }
+
+    @Override
+    public String toSql(CustomEnum val) {
+        return val != null ? val.name() : null;
+    }
+}