+++ jaorm-processor/src/test/resources/relationship/RelEntityCustomDAO.java
@@ -1,0 +1,7 @@
+package io.test;
+
+import io.github.ulisse1996.jaorm.BaseDao;
+import io.github.ulisse1996.jaorm.annotation.Dao;
+
+@Dao
+public interface RelEntityCustomDAO extends BaseDao<RelEntityCustom> {}