+++ jaorm-processor/src/test/resources/relationship/CustomEnum.java
@@ -1,0 +1,6 @@
+package io.test;
+
+public enum CustomEnum {
+    NAME1,
+    NAME2
+}