+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.1.1
+- Bugfix: Remove query load for cascade impl
+- Bugfix: Fix conversion on id columns
+
 ## 1.1.0
 - Feature: Automatic set of Relationship columns
 - Bugfix: Fix NPE on Listener Service for empty events